+++ source/host/desktop_session_agent.h
@@ -63,7 +63,7 @@ private slots:
     void onCursorPositionChanged(const QPoint& position);
     void onScreenTypeChanged(base::ScreenCapturer::ScreenType type, const QString& name);
     void onIpcDisconnected();
-    void onIpcMessageReceived(quint8 channel_id, const QByteArray& buffer);
+    void onIpcMessageReceived(quint32 channel_id, const QByteArray& buffer);
     void onClipboardEvent(const proto::desktop::ClipboardEvent& event);
 
 private:
