+++ source/host/desktop_session_ipc.cc
@@ -176,7 +176,7 @@ void DesktopSessionIpc::onIpcDisconnected()
 }
 
 //--------------------------------------------------------------------------------------------------
-void DesktopSessionIpc::onIpcMessageReceived(quint8 channel_id, const QByteArray& buffer)
+void DesktopSessionIpc::onIpcMessageReceived(quint32 channel_id, const QByteArray& buffer)
 {
     if (channel_id != proto::internal::CHANNEL_ID_SESSION)
     {
