+++ source/host/desktop_session_ipc.h
@@ -53,7 +53,7 @@ public:
 
 private slots:
     void onIpcDisconnected();
-    void onIpcMessageReceived(quint8 channel_id, const QByteArray& buffer);
+    void onIpcMessageReceived(quint32 channel_id, const QByteArray& buffer);
 
 private:
     using SharedBuffers = QMap<int, base::SharedPointer<base::SharedMemory>>;
