+++ source/host/user_session_agent.h
@@ -86,7 +86,7 @@ signals:
 
 private slots:
     void onIpcDisconnected();
-    void onIpcMessageReceived(quint8 channel_id, const QByteArray& buffer);
+    void onIpcMessageReceived(quint32 channel_id, const QByteArray& buffer);
 
 private:
     void sendSessionMessage();
